+++ packages/itensor/quri_parts/itensor/estimator.py
@@ -1,12 +1,10 @@
 import os
-from collections.abc import Iterable, Mapping
 from typing import NamedTuple, Union
 
 import juliacall
 from juliacall import Main as jl
-from quri_parts.circuit import NonParametricQuantumCircuit, gate_names
 from quri_parts.core.estimator import Estimatable, Estimate, QuantumEstimator
-from quri_parts.core.operator import Operator, PauliLabel, pauli_name, zero
+from quri_parts.core.operator import Operator, zero
 from quri_parts.core.state import (
     CircuitQuantumState,
     ParametricCircuitQuantumState,
@@ -15,12 +13,14 @@ from quri_parts.core.state import (
 )
 from typing_extensions import TypeAlias
 
+from .circuit import convert_circuit
+from .operator import convert_operator
+
 path = os.getcwd()
 library_path = os.path.join(path, "packages/itensor/quri_parts/itensor/library.jl")
 
 jl.seval("using ITensors")
 include_statement = 'include("' + library_path + '")'
-print(include_statement)
 jl.seval(include_statement)
 
 
@@ -38,96 +38,6 @@ QulacsStateT: TypeAlias = Union[CircuitQuantumState, QuantumStateVector]
 QulacsParametricStateT: TypeAlias = Union[
     ParametricCircuitQuantumState, ParametricQuantumStateVector
 ]
-
-
-from quri_parts.circuit.gate_names import (
-    MultiQubitGateNameType,
-    SingleQubitGateNameType,
-    TwoQubitGateNameType,
-    is_gate_name,
-    is_multi_qubit_gate_name,
-    is_parametric_gate_name,
-    is_single_qubit_gate_name,
-    is_two_qubit_gate_name,
-    is_unitary_matrix_gate_name,
-)
-
-# For now, we only support gates which is defined [here](https://github.com/ITensor/ITensors.jl/blob/d5ed4061f1e6224d0135fd7690c3be2fbecd0d9d/src/physics/site_types/qubit.jl)
-
-_single_qubit_gate_itensor: Mapping[SingleQubitGateNameType, str] = {
-    gate_names.X: "X",
-    gate_names.Y: "Y",
-    gate_names.Z: "Z",
-    gate_names.H: "H",
-    gate_names.S: "S",
-}
-
-_single_qubit_rotation_gate_itensor: Mapping[SingleQubitGateNameType, str] = {
-    gate_names.RX: "Rx",
-    gate_names.RY: "Ry",
-    gate_names.RZ: "Rz",
-}
-
-_two_qubit_gate_itensor: Mapping[TwoQubitGateNameType, str] = {
-    gate_names.CNOT: "CNOT",
-    gate_names.CZ: "CZ",
-    gate_names.SWAP: "SWAP",
-}
-
-
-def convert_circuit(
-    circuit: NonParametricQuantumCircuit, s: juliacall.VectorValue
-) -> juliacall.VectorValue:
-    gate_list: juliacall.VectorValue = jl.gate_list()
-    for gate in circuit.gates:
-        if not is_gate_name(gate.name):
-            raise ValueError(f"Unknown gate name: {gate.name}")
-
-        if is_single_qubit_gate_name(gate.name):
-            if gate.name in _single_qubit_gate_itensor:
-                gate_list = jl.add_gate(
-                    gate_list,
-                    _single_qubit_gate_itensor[gate.name],
-                    gate.target_indices[0] + 1,
-                )
-            elif gate.name in _single_qubit_rotation_gate_itensor:
-                gate_list = jl.add_gate(
-                    gate_list,
-                    _single_qubit_rotation_gate_itensor[gate.name],
-                    gate.target_indices[0] + 1,
-                    gate.params[0],
-                )
-            else:
-                raise ValueError(f"Unknown single qubit gate name: {gate.name}")
-        elif is_two_qubit_gate_name(gate.name):
-            gate_list = jl.add_gate(
-                gate_list,
-                _two_qubit_gate_itensor[gate.name],
-                gate.target_indices[0] + 1,
-                gate.target_indices[1] + 1,
-            )
-        else:
-            raise ValueError(f"Unknown gate name: {gate.name}")
-    circuit = jl.ops(gate_list, s)
-    return circuit
-
-
-def convert_operator(
-    operator: Union[Operator, PauliLabel], s: juliacall.VectorValue
-) -> juliacall.AnyValue:
-    paulis: Iterable[tuple[PauliLabel, complex]]
-    if isinstance(operator, Operator):
-        paulis = operator.items()
-    else:
-        paulis = [(operator, 1)]
-    os: juliacall.AnyValue = jl.OpSum()
-    for pauli, coef in paulis:
-        pauli_gates: juliacall.VectorValue = jl.gate_list()
-        for i, p in pauli:
-            pauli_gates = jl.add_pauli(pauli_gates, pauli_name(p), i + 1)
-        os = jl.add_coef_pauli(os, coef, pauli_gates)
-    op: juliacall.AnyValue = jl.MPO(os, s)
-    return op
 
 
 def _estimate(operator: Estimatable, state: QulacsStateT) -> Estimate[complex]:
